+++ src/test/kotlin/com/github/cramer/gradle/jasperreports/MetaTest.kt
@@ -3,6 +3,7 @@ package com.github.cramer.gradle.jasperreports
 import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNotEmpty
+import com.github.fcramer.gradle.jasperreports.utils.getJasperreportsVersion
 import org.junit.jupiter.api.Test
 
 class MetaTest {